fix: respect manifold dtype under global jax_enable_x64 via boundary casts
When jax_enable_x64 is enabled globally, array constructors without explicit
dtype= (weights via jax.random.*, jnp.zeros/full/arange) become float64. This
previously leaked into computations before manifold methods could _cast them,
silently promoting float32 manifolds to float64 and breaking the dtype contract.

Apply boundary casts at points where float64-valued params combine with
float32 inputs in raw Euclidean ops (einsum/matmul/@), before reaching a
_cast-ing manifold method:

- hyperboloid_linear: cast kernel/bias in _fhcnn/_fhnn forward paths and HTCLinear
- hyperboloid_positional: build rotary arange in input dtype to avoid int64→float64
- poincare_linear: cast kernel in tangent-space einsum
- poincare_batchnorm: pin learned/running stats to manifold.dtype at init
- hyperboloid_regularization: cast bias/running_mean at use, add dtype= ctor param

Add regression test under conftest's global x64: float32 manifold + layer +
input must yield float32 output (and float32 persistent state for BatchNorms).

+++ hyperbolix/nn_layers/hyperboloid_linear.py
@@ -56,7 +56,11 @@ def _fhcnn_forward(
     if activation is not None:
         x_BI = activation(x_BI)
 
-    # Linear transformation: (B, in_dim) -> (B, out_dim)
+    # Linear transformation: (B, in_dim) -> (B, out_dim). Cast params to the
+    # input dtype so float64 weights (created when jax_enable_x64 is enabled
+    # globally) don't silently promote a float32 manifold computation.
+    kernel_OI = kernel_OI.astype(x_BI.dtype)
+    bias_1O = bias_1O.astype(x_BI.dtype)
     x_BO = jnp.einsum("bi,oi->bo", x_BI, kernel_OI) + bias_1O
 
     # Split into time and space: x0 is first coord, x_rem is spatial
@@ -121,7 +125,11 @@ def _fhnn_forward(
     if dropout_fn is not None:
         x_BI = dropout_fn(x_BI)
 
-    # Linear transformation: (B, in_dim) -> (B, out_dim)
+    # Linear transformation: (B, in_dim) -> (B, out_dim). Cast params to the
+    # input dtype so float64 weights (created when jax_enable_x64 is enabled
+    # globally) don't silently promote a float32 manifold computation.
+    kernel_OI = kernel_OI.astype(x_BI.dtype)
+    bias_1O = bias_1O.astype(x_BI.dtype)
     z_BO = jnp.einsum("bi,oi->bo", x_BI, kernel_OI) + bias_1O  # (B, O)
 
     # Split into time logit and spatial components
@@ -731,9 +739,11 @@ class HTCLinear(nnx.Module):
         """
 
         def linear_fn(z):
-            out = z @ self.kernel[...]
+            # Cast params to the working dtype so float64 weights (from global
+            # jax_enable_x64) don't promote a float32 computation to float64.
+            out = z @ self.kernel[...].astype(z.dtype)
             if self.bias is not None:
-                out = out + self.bias[...]
+                out = out + self.bias[...].astype(z.dtype)
             return out
 
         return htc(x, linear_fn, c_in, c_out, self.eps)
